Update master branch to the Q2 release.

+++ data-warehouse-postgresql/gobii_ifl/gobii_ifl/gobii_ifl.py
@@ -6,8 +6,12 @@ import os
 import getopt
 import load_ifile
 import preprocess_ifile
+import deduplicate_ifile
 from util.ifl_utility import IFLUtility
+from os.path import basename
+from os.path import splitext
 
+#exit codes used here: 1, 2, 3, 4
 def main(argv):
 		verbose = False
 		flCheck = False
@@ -43,32 +47,89 @@ def main(argv):
 		#print("Opts: ", opts)
 		if connectionStr != "" and outputPath != "":
 			if inputDir != "":
+				#Directory iterative run
 				for f in os.listdir(inputDir):
 					try:
 						if verbose:
 							print("Processing file %s..." % f)
-						preprocessedFile = preprocess_ifile.main(verbose, connectionStr, os.path.join(inputDir, f), outputPath)
-						if flCheck and not checkDataIntegrity(f, preprocessedFile, verbose):
-							IFLUtility.printError("File length mismatch detected on %s. You have duplicate entries in the table where the NMAP file maps to, please fix it. Skipping file." % f)
-						else:
-							loadFile = load_ifile.main(verbose, connectionStr, preprocessedFile, outputPath)
+						tableName = splitext(basename(f))[1][1:]
+						isProp = tableName.endswith('_prop')
+						if isProp:
+							longPropFilename = outputPath+"long_"+basename(f)
+						#preprocessing happens here
+						preprocessedFile, exitCode = preprocess_ifile.main(verbose, connectionStr, os.path.join(inputDir, f), outputPath)
+
+						if exitCode != 0:
+							sys.exit(exitCode)
+
+						#use the input file vs ppd file for general cases
+						if flCheck and not isProp:
+							if not checkDataIntegrity(f, preprocessedFile, verbose):
+								IFLUtility.printError("File length mismatch detected on %s. You either have duplicate entries in the table where the NMAP file maps to OR you're trying to load data to entities that do not exist, please fix it first. Loading will abort." % f)
+								exitCode = 4
+								sys.exit(exitCode)
+						#use the long format file vs the ppd file for property data
+						if flCheck and isProp:
+							if not checkDataIntegrity(longPropFilename, preprocessedFile, verbose):
+								IFLUtility.printError("File length mismatch detected on %s. You either have duplicate entries in the table where the NMAP file maps to OR you're trying to load data to entities that do not exist, please fix it first. Loading will abort." % f)
+								exitCode = 5
+								sys.exit(exitCode)
+
+						if not isProp:
+							dedupFile, exitCode = deduplicate_ifile.main(verbose, preprocessedFile, outputPath, tableName)
+							loadFile, exitCode = load_ifile.main(verbose, connectionStr, dedupFile, outputPath)
+						else:  # DO NOT DEDUPLICATE PROPERTY FILES AS YOU'LL REMOVE GOOD DATA, NO NEED FOR IT ANYWAY SINCE THIS WILL BE AN UPSERT OPERATION
+							loadFile, exitCode = load_ifile.main(verbose, connectionStr, preprocessedFile, outputPath)
+						if exitCode != 0:
+							sys.exit(exitCode)
+						'''
+						In light of enhanced error logging, we decided that there is value to these files and so IFLs will not delete them.
 						try:
 							os.remove(preprocessedFile)
 							os.remove(loadFile)
 						except Exception as e:
 							IFLUtility.printError("Failed to remove temporary files. Check file permissions. Error: %s" % str(e))
 							exitCode = 2
+						'''
 					except Exception as e1:
 						IFLUtility.printError("Failed to load file %s. Error: %s" % (f, str(e1)))
 						exitCode = 3
+						sys.exit(exitCode)
 			elif iFile != "":
-				preprocessedFile = preprocess_ifile.main(verbose, connectionStr, iFile, outputPath)
-				#return
+				#Per file run
+				tableName = splitext(basename(iFile))[1][1:]
+				isProp = tableName.endswith('_prop')
+				if isProp:
+					longPropFilename = outputPath+"long_"+basename(iFile)
+				#preprocessing happens here
+				preprocessedFile, exitCode = preprocess_ifile.main(verbose, connectionStr, iFile, outputPath)
+
+				#terminate if there are any errors in preprocessing stage
+				if exitCode != 0:
+					sys.exit(exitCode)
 				loadFile = None
-				if flCheck and not checkDataIntegrity(iFile, preprocessedFile, verbose):
-					IFLUtility.printError("File length mismatch detected on %s. You have duplicate entries in the table where the NMAP file maps to, please fix it first. Loading will abort." % iFile)
-				else:
-					loadFile = load_ifile.main(verbose, connectionStr, preprocessedFile, outputPath)
+				#use the input file vs ppd file for general cases
+				if flCheck and not isProp:
+					if not checkDataIntegrity(iFile, preprocessedFile, verbose):
+						IFLUtility.printError("File length mismatch detected on %s. You either have duplicate entries in the table where the NMAP file maps to OR you're trying to load data to entities that do not exist, please fix it first. Loading will abort." % iFile)
+						exitCode = 4
+						sys.exit(exitCode)
+				#use the long format file vs the ppd file for property data
+				if flCheck and isProp:
+					if not checkDataIntegrity(longPropFilename, preprocessedFile, verbose):
+						IFLUtility.printError("File length mismatch detected on %s. You either have duplicate entries in the table where the NMAP file maps to OR you're trying to load data to entities that do not exist, please fix it first. Loading will abort." % iFile)
+						exitCode = 5
+						sys.exit(exitCode)
+				if not isProp:
+					dedupFile, exitCode = deduplicate_ifile.main(verbose, preprocessedFile, outputPath, tableName)
+					loadFile, exitCode = load_ifile.main(verbose, connectionStr, dedupFile, outputPath)
+				else:  # DO NOT DEDUPLICATE PROPERTY FILES AS YOU'LL REMOVE GOOD DATA, NO NEED FOR IT ANYWAY SINCE THIS WILL BE AN UPSERT OPERATION
+					loadFile, exitCode = load_ifile.main(verbose, connectionStr, preprocessedFile, outputPath)
+
+				if exitCode != 0:
+						sys.exit(exitCode)
+				'''
+				In light of enhanced error logging, we decided that there is value to these files and so IFLs will not delete them.
 				try:
 					os.remove(preprocessedFile)
 					os.remove(loadFile)
@@ -76,10 +137,12 @@ def main(argv):
 				except Exception as e:
 					IFLUtility.printError("Failed to remove temporary files. Check file permissions. Error: %s" % str(e))
 					exitCode = 0
+				'''
 			else:
 				printUsageHelp()
 		else:
 			printUsageHelp()
+		#return rowsLoaded
 		sys.exit(exitCode)
 		#cleanup
 
@@ -105,6 +168,7 @@ def printUsageHelp():
 	print ("\t-v or --verbose = Print the status of the IFL in more detail")
 	print ("\t-l or --fileLengthCheck = This will check if the preprocessed file is of the same length as the input file. \n\t\tA mismatch indicates duplicate entries in the table where the NMAP file maps to.\n\t\tInput file should not be loaded in that case.")
 	sys.exit(1)
+
 
 if __name__ == "__main__":
 	main(sys.argv[1:])
