+++ gobiiproject/gobii-dtomapping/src/main/java/org/gobiiproject/gobiidtomapping/impl/DtoMapCvImpl.java
@@ -1,13 +1,14 @@
 package org.gobiiproject.gobiidtomapping.impl;
 
-import org.gobiiproject.gobiidao.GobiiDaoException;
 import org.gobiiproject.gobiidao.resultset.access.RsCvDao;
 import org.gobiiproject.gobiidao.resultset.core.ParamExtractor;
 import org.gobiiproject.gobiidao.resultset.core.ResultColumnApplicator;
 import org.gobiiproject.gobiidtomapping.DtoMapCv;
+import org.gobiiproject.gobiidtomapping.DtoMapCvGroup;
 import org.gobiiproject.gobiidtomapping.GobiiDtoMappingException;
-import org.gobiiproject.gobiimodel.dto.container.CvDTO;
-import org.gobiiproject.gobiimodel.entity.CvItem;
+import org.gobiiproject.gobiimodel.headerlesscontainer.CvDTO;
+import org.gobiiproject.gobiimodel.headerlesscontainer.CvGroupDTO;
+import org.gobiiproject.gobiimodel.types.GobiiCvGroupType;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,6 +16,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.ArrayList;
+import java.util.List;
 import java.util.Map;
 
 /**
@@ -28,44 +30,50 @@ public class DtoMapCvImpl implements DtoMapCv {
     @Autowired
     private RsCvDao rsCvDao = null;
 
+    @Autowired
+    DtoMapCvGroup dtoMapCvGroup;
+
+
     @Override
-    public CvDTO getCvDetails(CvDTO cvDTO) throws GobiiDtoMappingException {
+    public List<CvDTO> getCvs() throws GobiiDtoMappingException {
+
+        List<CvDTO> returnVal = new ArrayList<>();
+
+        try {
+            ResultSet resultSet = rsCvDao.getCvNames();
+            while (resultSet.next()) {
+                CvDTO currentCvDTO = new CvDTO();
+                currentCvDTO.setTerm(resultSet.getString("term"));
+                currentCvDTO.setCvId(resultSet.getInt("cv_id"));
+                currentCvDTO.setGroupType(resultSet.getInt("group_type"));
+                returnVal.add(currentCvDTO);
+            }
+        } catch (SQLException e) {
+            LOGGER.error("Gobii Mapping error", e);
+            throw new GobiiDtoMappingException(e);
+        }
+
+        return returnVal;
+    }
+
+    @Override
+    public CvDTO getCvDetails(Integer cvId) throws GobiiDtoMappingException {
 
         CvDTO returnVal = new CvDTO();
 
         try {
 
-            ResultSet resultSet = rsCvDao.getDetailsForCvId(cvDTO.getCvId());
+            ResultSet resultSet = rsCvDao.getDetailsForCvId(cvId);
 
             if (resultSet.next()) {
                 // apply cv values
                 ResultColumnApplicator.applyColumnValues(resultSet, returnVal);
+
             }
 
-            if (cvDTO.isIncludeDetailsList()) {
-                ResultSet cvItemsResultSet = rsCvDao.getAllCvItems();
-                String currentGroupName = "";
-                while (cvItemsResultSet.next()) {
-
-                    String newGroupName = cvItemsResultSet.getString("group");
-
-                    if (!currentGroupName.equals(newGroupName)) {
-                        currentGroupName = newGroupName; //set Group name if first Group name frm query
-                        returnVal.getGroupCvItems().put(currentGroupName, new ArrayList<>());
-                    }
-
-                    CvItem currentGroupCvItem = new CvItem();
-                    currentGroupCvItem.setCvId(cvItemsResultSet.getInt("cv_id"));
-                    currentGroupCvItem.setRank(cvItemsResultSet.getInt("rank"));
-                    currentGroupCvItem.setTerm(cvItemsResultSet.getString("term"));
-                    currentGroupCvItem.setDefinition(cvItemsResultSet.getString("definition"));
-                    returnVal.getGroupCvItems().get(currentGroupName).add(currentGroupCvItem);
-
-                }
-            }
-        } catch (Exception e) {
-            returnVal.getDtoHeaderResponse().addException(e);
-            LOGGER.error("Gobii Maping Error", e);
+        } catch (SQLException e) {
+            LOGGER.error("Gobii Mapping Error", e);
+            throw new GobiiDtoMappingException(e);
         }
 
 
@@ -77,33 +85,96 @@ public class DtoMapCvImpl implements DtoMapCv {
     public CvDTO createCv(CvDTO cvDTO) throws GobiiDtoMappingException {
         CvDTO returnVal = cvDTO;
 
-        try {
+        Integer groupIdForUserGroup = this.getUserCvGroupId(cvDTO);
 
-            Map<String, Object> parameters = ParamExtractor.makeParamVals(cvDTO);
+        if (groupIdForUserGroup > 0 ) {
+
+            returnVal.setGroup(groupIdForUserGroup);
+            Map<String, Object> parameters = ParamExtractor.makeParamVals(returnVal);
             Integer cvId = rsCvDao.createCv(parameters);
             returnVal.setCvId(cvId);
 
-        } catch (Exception e) {
-            returnVal.getDtoHeaderResponse().addException(e);
-            LOGGER.error("Gobii Maping Error", e);
+        } else {
+
+            String message = "There is no user cvgroup for the specified CV group";
+            if( cvDTO != null && cvDTO.getGroupId() != null ) {
+                message += ": " + cvDTO.getGroupId().toString();
+            }
+
+            throw new GobiiDtoMappingException(message);
         }
 
         return returnVal;
     }
 
+    private Integer getUserCvGroupId(CvDTO cvDto) throws GobiiDtoMappingException {
+
+        Integer returnVal;
+
+        if (cvDto != null) {
+
+            if (cvDto.getGroupId() != null && cvDto.getGroupId() > 0) {
+
+                CvGroupDTO cvGroupDTOCandidate = this.dtoMapCvGroup.getCvGroup(cvDto.getGroupId());
+                if (cvGroupDTOCandidate.getGroupType().equals(GobiiCvGroupType.GROUP_TYPE_USER.getGroupTypeId())) {
+
+                    returnVal = cvGroupDTOCandidate.getCvGroupId();
+
+                } else {
+
+                    CvGroupDTO cvGroupDTOUser = this.dtoMapCvGroup.getUserCvByGroupName(cvGroupDTOCandidate.getName());
+                    returnVal = cvGroupDTOUser.getCvGroupId();
+                }
+
+            } else {
+                throw new GobiiDtoMappingException("The input CvDTO does not have a group id");
+            }
+
+        } else {
+            throw new GobiiDtoMappingException("The input CvDTO is null");
+        }
+
+        return returnVal;
+    }
+
+
     @Override
-    public CvDTO updateCv(CvDTO cvDTO) throws GobiiDtoMappingException {
+    public CvDTO replaceCv(Integer cvId, CvDTO cvDTO) throws GobiiDtoMappingException {
 
         CvDTO returnVal = cvDTO;
 
-        try {
+        Integer groupIdForUserGroup = this.getUserCvGroupId(cvDTO);
 
-            Map<String, Object> parameters = ParamExtractor.makeParamVals(returnVal);
-            rsCvDao.updateCv(parameters);
+        if (groupIdForUserGroup > 0 ) {
 
-        } catch (Exception e) {
-            returnVal.getDtoHeaderResponse().addException(e);
-            LOGGER.error("Gobii Maping Error", e);
+            CvDTO currentCvDTO = getCvDetails(cvId);
+
+            if (currentCvDTO.getCvId() > 0 ) {
+
+                cvDTO.setGroup(groupIdForUserGroup);
+                cvDTO.setCvId(cvId);
+                Map<String, Object> parameters = ParamExtractor.makeParamVals(returnVal);
+                rsCvDao.updateCv(parameters);
+
+            } else {
+
+                String message = "There is no cv for id";
+                if( cvDTO != null  ) {
+                    message += ": " + cvId.toString();
+                }
+
+                LOGGER.error(message);
+                throw new GobiiDtoMappingException(message);
+            }
+
+        } else {
+
+            String message = "There is no user cvgroup for the specified CV group";
+            if( cvDTO != null && cvDTO.getGroupId() != null  ) {
+                message += ": " + cvDTO.getGroupId().toString();
+            }
+            throw new GobiiDtoMappingException(message);
+
         }
 
         return returnVal;
@@ -114,16 +185,62 @@ public class DtoMapCvImpl implements DtoMapCv {
 
         CvDTO returnVal = cvDTO;
 
-        try {
+        if (cvDTO.getGroupType().equals(GobiiCvGroupType.GROUP_TYPE_USER.getGroupTypeId())) {
 
+            returnVal.setEntityStatus(0);
             Map<String, Object> parameters = ParamExtractor.makeParamVals(returnVal);
             rsCvDao.deleteCv(parameters);
 
-        } catch (Exception e) {
-            returnVal.getDtoHeaderResponse().addException(e);
-            LOGGER.error("Gobii Maping Error", e);
+            returnVal.setCvId(-1);
+            returnVal.setGroupId(null);
+            returnVal.setXrefId(null);
+            returnVal.setTerm(null);
+            returnVal.setAbbreviation(null);
+            returnVal.setDefinition(null);
+            returnVal.setRank(null);
+
+        } else {
+
+            String message = "Cannot delete cv term that belongs to a system group";
+            if( cvDTO != null && cvDTO.getCvId() != null ) {
+                message += ": " + cvDTO.getCvId().toString();
+            }
+
+            LOGGER.error(message);
+            throw new GobiiDtoMappingException(message);
+
         }
 
         return returnVal;
     }
+
+    @Override
+    public List<CvDTO> getCvsByGroupName(String groupName) throws GobiiDtoMappingException {
+
+        List<CvDTO> returnVal = new ArrayList<>();
+
+        try {
+
+            ResultSet resultSet = rsCvDao.getCvsByGroup(groupName);
+
+            while (resultSet.next()) {
+
+                CvDTO currentCvDTO = new CvDTO();
+
+                ResultColumnApplicator.applyColumnValues(resultSet, currentCvDTO);
+                returnVal.add(currentCvDTO);
+
+            }
+
+
+        } catch (SQLException e) {
+
+            LOGGER.error("Gobii Mapping Error", e);
+            throw new GobiiDtoMappingException(e);
+        }
+
+        return returnVal;
+
+    }
+
 } // DtoMapNameIdListImpl
